Rename for consistency.
Controller -> Controllers

Related to: #8
